Add fdisk and azcopy packages and fix writing bashrc each time entering the container

+++ advrisc/u24.04-qcsubv1/entrypoint.sh
@@ -24,16 +24,24 @@ if ! grep -q "^$UNAME " /etc/sudoers; then
     echo "$UNAME ALL=(ALL) NOPASSWD: ALL" >> /etc/sudoers
 fi
 
-cp -r --update=none /etc/skel/. "/home/$UNAME"/
+INITIALIZED_FLAG="/.dev_env_initialized"
+
+if [ ! -f "$INITIALIZED_FLAG" ]; then
+    echo "First time setup for user $UNAME..."
+    cp -r --update=none /etc/skel/. "/home/$UNAME"/
+
+    BASHRC="/home/$UNAME/.bashrc"
+    echo "export HOME=/home/$UNAME" >> "$BASHRC"
+    echo "export USER=$UNAME" >> "$BASHRC"
+
+    chown "$HOST_UID:$HOST_GID" "/home/$UNAME"
+    chown "$HOST_UID:$HOST_GID" "/home/$UNAME/".*
+
+    touch "$INITIALIZED_FLAG"
+fi
 
 export HOME=/home/$UNAME
 export USER=$UNAME
-
-BASHRC="/home/$UNAME/.bashrc"
-echo "export HOME=/home/$UNAME" >> "$BASHRC"
-echo "export USER=$UNAME" >> "$BASHRC"
-
-chown "$HOST_UID:$HOST_GID" "/home/$UNAME/".*
 
 cd $HOME
 exec gosu "$UNAME" "$@"
